+++ src/index.js
@@ -1,6 +1,6 @@
 import neuralNetwork from "./NeuralNetwork";
 import handpose from "./Handpose";
-import poseDetection from "./PoseDetection";
+import bodypose from "./Bodypose";
 import * as tf from "@tensorflow/tfjs";
 import * as tfvis from "@tensorflow/tfjs-vis";
 import p5Utils from "./utils/p5Utils";
@@ -13,7 +13,7 @@ export default Object.assign(
     tfvis,
     neuralNetwork,
     handpose,
-    poseDetection,
+    bodypose,
     setBackend,
   }
 );
